libril: Schedule release as close as possible.
In the current implementation, release of wake lock is scheduled one second
in the future. If there are other pending events in the queue, it will be
rescheduled one second after the last one. This causes one second of extra
awake time, even after handling all events. On my phone, this amounts to about
20-30 minutes a day.
The correct way to implement this would to to add a function to
ril_event.{h,cpp} that allows appending an event to the end of the list,
without specifying a hard timeout. Unfortunately, the rescheduling logic will
have to stay, as more events could have been added to the list in the mean
time.
My submitted change does reduce the extra time held awake by about 5x. A
more complete implementation that adds the release of the wake lock to the end
of the list will only save a couple of reshedules, but won't change much
functionality-wise.

Change-Id: Ie319a7837818d8e971fb436a192f6ce98e699944